+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * throw exception when more unquote then quasiquotes
 * fix double unquote symbol
 * throw error when unquote-splice on atom in the middle of the list
+* don't create cycles in unquote-splicing
 
 ## 0.15.1
 ### Bug fixes